# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BuntesBegegnen.Api",
    "version": "1.0"
  },
  "paths": {
    "/account/register": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequestDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationErrorDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationErrorDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationErrorDto"
                }
              }
            }
          }
        }
      }
    },
    "/account/login": {
      "post": {
        "tags": [
          "Account"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "lastName": {
            "maxLength": 25,
            "minLength": 2,
            "type": "string"
          },
          "birthDay": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "email": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string"
          },
          "phoneNumber": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "streetName": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "houseNumber": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "zipCode": {
            "maxLength": 5,
            "minLength": 0,
            "pattern": "\\d{5}",
            "type": "string",
            "nullable": true
          },
          "city": {
            "maxLength": 100,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "goals": {
            "maxLength": 10000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "hobbies": {
            "maxLength": 10000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "impairedSight": {
            "type": "boolean"
          },
          "impairedHearing": {
            "type": "boolean"
          },
          "impairedSpeech": {
            "type": "boolean"
          },
          "impairedMobility": {
            "type": "boolean"
          },
          "additionalHandicaps": {
            "maxLength": 5000,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "LoginRequestDto": {
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "RegistrationErrorDto": {
        "required": [
          "errors"
        ],
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "additionalProperties": false
      },
      "RegistrationRequestDto": {
        "required": [
          "account",
          "password",
          "previewModePassword"
        ],
        "type": "object",
        "properties": {
          "account": {
            "$ref": "#/components/schemas/CreateUserDto"
          },
          "password": {
            "type": "string"
          },
          "previewModePassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter the authentication token (without the 'Bearer' prefix)",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}